Finish a3 p2

+++ COMP7404/a3/parse.py
@@ -46,7 +46,7 @@ def read_grid_mdp_problem_p2(file_path):
         file[x] = line.strip("\n")
 
     res = {
-        "discount": int(file[0].replace("discount: ", "")),
+        "discount": float(file[0].replace("discount: ", "")),
         "noise": float(file[1].replace("noise: ", "")),
         "livingReward": float(file[2].replace("livingReward: ", "")),
         "iterations": int(file[3].replace("iterations: ", "")),
